Small fixes and large pack pack temp solution
