+++ crawlers/txpool/sample.env
@@ -1,5 +1,7 @@
 export MOONSTREAM_NODE_ETHEREUM_IPC_ADDR="127.0.0.1"
 export MOONSTREAM_NODE_ETHEREUM_IPC_PORT="8545"
-export ETHTXPOOL_HUMBUG_CLIENT_ID="<client id for the crawling machine>"
-export ETHTXPOOL_HUMBUG_TOKEN="<Generate an integration and a Humbug token from https://bugout.dev/account/teams>"
+export MOONSTREAM_NODE_POLYGON_IPC_ADDR="127.0.0.1"
+export MOONSTREAM_NODE_POLYGON_IPC_PORT="8545"
+export HUMBUG_TXPOOL_CLIENT_ID="<client id for the crawling machine>"
+export HUMBUG_TXPOOL_TOKEN="<Generate an integration and a Humbug token from https://bugout.dev/account/teams>"
 export HUMBUG_REPORTER_CRAWLERS_TOKEN="<Bugout Humbug token for crash reports>"
